Moved Product Settings Sections and Rows in two separate files
