+++ WooCommerce/Classes/ViewRelated/Products/Edit Product/Product Settings/ProductSettingsSections.swift
@@ -1,0 +1,38 @@
+import UIKit
+import Yosemite
+
+// MARK: - Mediators Protocols
+//
+/// Encapsulates configuration and interaction of a UITableView section header.
+protocol ProductSettingsSectionMediator {
+    var title: String { get }
+    var rows: [ProductSettingsRowMediator] { get }
+
+    init(_ product: Product)
+}
+
+// MARK: - Sections declaration for Product Settings
+//
+enum ProductSettingsSections {
+    /// Publish Settings section
+    struct PublishSettings: ProductSettingsSectionMediator {
+        let title = NSLocalizedString("Publish Settings", comment: "Title of the Publish Settings section on Product Settings screen")
+
+        let rows: [ProductSettingsRowMediator]
+
+        init(_ product: Product) {
+            rows = [ProductSettingsRows.CatalogVisibility(product)]
+        }
+    }
+
+    /// More Settings section
+    struct MoreOptions: ProductSettingsSectionMediator {
+        let title = NSLocalizedString("More Options", comment: "Title of the More Options section on Product Settings screen")
+
+        let rows: [ProductSettingsRowMediator]
+
+        init(_ product: Product) {
+            rows = [ProductSettingsRows.Slug(product)]
+        }
+    }
+}
